use cascaded deletes in wipeout
